changed fov to float

+++ include/depthai-shared/common/CameraInfo.hpp
@@ -9,7 +9,7 @@ struct CameraInfo {
     std::vector<std::vector<float>> intrinsicMatrix;
     std::vector<float> distortionCoeff;
     Extrinsics extrinsics;
-    double measuredFovDeg;  // fov in deg
+    float measuredFovDeg;  // fov in deg
     // TODO(sachin): Should I add type of distortion model here ?
     NLOHMANN_DEFINE_TYPE_INTRUSIVE(CameraInfo, intrinsicMatrix, width, height, distortionCoeff, extrinsics, measuredFovDeg);
 };
